Add user preferred device
Add to the users table an attribute for storing the user's preferred
device that will be shown on the dashboard. Added an option to the user
form for setting a preferred device and displaying it on the user show
page. Lastly set the dashboard to display the page the user's preferred
device is on and have it selected to start with.