Add device status to dashboard
Add text underneath the device name on the device table that says if the
device is online or when it was last seen. Add a new attribute to the
device table that saves the datetime of when the device last sent an
update API call. The dashboard uses this data to check if the device has
missed three updates based on its update_rate plus an extra minute to
mark the device as stale. Add lock to changing the selected device to
prevent a race condition.