Add stale image alert to dashboard
Add an closeable warning alert to the bottom of the device image for
when the devices image is stale. The dashboard uses the images
updated_at value to check if the device has missed three image updates
based on its image_rate plus an extra minute to mark the image as stale.